🚚 Rename & split files
1. Renamed `LuminareButtonCompose` to `LuminareButton`.
2. Formatted some closures.
3. Renamed folder `Compose` to `Composes`, as it represents a flattened structure of composed components.
4. Splitted `LuminareButtonStyles` into multiple files.
5. Renamed and collected some view modifiers.
