+++ Sources/Luminare/Components/Button Styles/LuminareButtonStyle.swift
@@ -1,0 +1,52 @@
+//
+//  LuminareButtonStyle.swift
+//  Luminare
+//
+//  Created by KrLite on 2025/4/12.
+//
+
+import SwiftUI
+
+/// A stylized button style.
+///
+/// ![LuminareButtonStyle](LuminareButtonStyle)
+public struct LuminareButtonStyle: ButtonStyle {
+    @Environment(\.isEnabled) private var isEnabled
+    @Environment(\.luminareAnimationFast) private var animationFast
+    @Environment(\.luminareMinHeight) private var minHeight
+    @Environment(\.luminareButtonMaterial) private var material
+    @Environment(\.luminareButtonCornerRadii) private var cornerRadii
+    @Environment(\.luminareButtonHighlightOnHover) private var highlightOnHover
+
+    @State private var isHovering: Bool
+
+    public init() {
+        self.isHovering = false
+    }
+
+    #if DEBUG
+        init(
+            isHovering: Bool = false
+        ) {
+            self.isHovering = isHovering
+        }
+    #endif
+
+    public func makeBody(configuration: Configuration) -> some View {
+        configuration.label
+            .frame(maxWidth: .infinity, maxHeight: .infinity)
+            .onHover { hover in
+                withAnimation(animationFast) {
+                    isHovering = hover
+                }
+            }
+            .frame(minHeight: minHeight)
+            .opacity(isEnabled ? 1 : 0.5)
+            .modifier(LuminareFilledModifier(
+                isHovering: isHovering, isPressed: configuration.isPressed,
+                fill: .quinary, hovering: .quaternary.opacity(0.7),
+                pressed: .quaternary
+            ))
+            .clipShape(.rect(cornerRadii: cornerRadii))
+    }
+}
