+++ Sources/Luminare/Components/Button Styles/LuminareCompactButtonStyle.swift
@@ -1,0 +1,51 @@
+//
+//  LuminareCompactButtonStyle.swift
+//  Luminare
+//
+//  Created by KrLite on 2025/4/12.
+//
+
+import SwiftUI
+
+/// A stylized button style with a border.
+///
+/// Can be configured to disable padding when `extraCompact` is set to `true`.
+///
+/// ![LuminareCompactButtonStyle](LuminareCompactButtonStyle)
+public struct LuminareCompactButtonStyle: ButtonStyle {
+    @Environment(\.isEnabled) private var isEnabled
+    @Environment(\.luminareAnimationFast) private var animationFast
+    @Environment(\.luminareHorizontalPadding) private var horizontalPadding
+
+    @State private var isHovering: Bool
+
+    public init() {
+        self.isHovering = false
+    }
+
+    #if DEBUG
+        init(
+            isHovering: Bool = false
+        ) {
+            self.isHovering = isHovering
+        }
+    #endif
+
+    public func makeBody(configuration: Configuration) -> some View {
+        configuration.label
+            .padding(.horizontal, horizontalPadding)
+            .modifier(LuminareAspectRatioModifier())
+            .opacity(isEnabled ? 1 : 0.5)
+            .modifier(LuminareFilledModifier(
+                isHovering: isHovering, isPressed: configuration.isPressed,
+                fill: .quinary, hovering: .quaternary.opacity(0.7),
+                pressed: .quaternary
+            ))
+            .modifier(LuminareBorderedModifier(isHovering: isHovering))
+            .onHover { hover in
+                withAnimation(animationFast) {
+                    isHovering = hover
+                }
+            }
+    }
+}
